refactor: Add getFixture to CaseCtx for cases that share expensive setup
Adds a Fixture<T> = { key, setup(ctx) } type and ctx.getFixture(fixture),
memoized once per registerCases() call (i.e. once per transport dimension),
regardless of how many cases in the array ask for it or in what order. This
lets several cases share one expensive setup step without a vitest beforeAll,
which can't express "shared by only some cases in this flat array" and forces
that whole group out of the Case model.

Uses it to fold storage's 13 cases that shared a beforeAll-seeded Actor run
(previously stuck in tests/integration/cases/storage_grouped.cases.ts — not
published, not critical-eligible) into ordinary Case objects in storage.cases.ts,
via a normalModeRunFixture + withNormalModeRun helper. All 18 storage cases
are now uniform, flat, and critical-eligible like every other group; the
setup Actor run still only executes once per dimension.

- src/test_kit/types.ts: Fixture<T>, CaseCtx.getFixture.
- src/test_kit/register.ts: builds one fixture cache per describe() block,
  injects getFixture into the ctx passed to every case's run().
- src/test_kit/cases/storage.cases.ts: merges storage_grouped's 13 cases in,
  now 18 total (was 5 + 13 local-only).
- Deletes tests/integration/cases/storage_grouped.cases.ts and its call site
  in suite.ts.
- Docs (AGENTS.md, DEVELOPMENT.md) updated to describe getFixture instead of
  the "one group doesn't fit the model" caveat.

Verified: type-check, lint, format, unit tests (91 files/1267 tests), build,
check:agents, and integration-suite collection count (117 tests per entry
file, 353 total — unchanged from before this change, confirming the 13
re-homed cases collect 1:1, zero loss).
