+++ src/test_kit/index.ts
@@ -12,7 +12,7 @@
 export { createMcpStatelessClient, createMcpStreamableClient } from './mcp_client.js';
 export type { SuiteClientOptions } from './mcp_client.js';
 export { registerCases } from './register.js';
-export type { Case, CaseCtx, SuiteClient, Transport } from './types.js';
+export type { Case, CaseCtx, Fixture, SuiteClient, Transport } from './types.js';
 
 export { actorsCases } from './cases/actors.cases.js';
 export { appsCases } from './cases/apps.cases.js';
